feat(inline_markdown): add split nodes delimiter function
It takes a list of nodes, a delimiter and a new text type.
It splits using the delimiter and parses the inner values from the
result of the split into the new text_type;
If one of the nodes is already not text_type it probably means that it
has already been delimited correctly.