feat(block_markdown): add validation for code type
The Regex verifies if the string ends and starts with backticks (`)
Then it verifies if there are 0 to any number of whitespaces or non
whitespaces, there must also be at least ONE non whitespace on the
string and then it continues with the same check.
This allows for ignoring strings with just whitespaces.